% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DID_IV_decomposition.R
\name{print_summary}
\alias{print_summary}
\title{Print the summary.}
\usage{
print_summary(data, return_df = FALSE)
}
\arguments{
\item{data}{A data.frame.}

\item{return_df}{Logical. If TRUE, returns the summary data.frame.}
}
\value{
Invisibly prints the summary to console. Returns a data.frame if return_df = TRUE.
}
\description{
Print the summary.
}
